recognizer/requirements.txt
python-dotenv
opencv-python
torch
Pillow
openai
transformers